Add files via upload
Write a C or C++ program that finds a practical measure of your machine’s SP (32 bit) and DP (64 bit)
precision by taking the difference of 2 numbers and comparing the result to zero in each data type. You
will submit the code, and in the written homework will state the values you obtained running your code
for each data type. You may find https://www.w3schools.com/cpp/cpp_while_loop.asp useful if you
are beginning in C++ programming.Use while and iterate on j = 0, 1, . . . . Hint: while ( ( 1−(1+ 1
2
j ) ̸= 0
) {. . . j++;}

Write a C or C++ program to multiply the integers 200*300*400*500 on your computer? What is the
result? You will submit the code, and in the written homework will name the effect you observed AND
provide a definition and math formula for underflow and overflow for IEEE SP and DP representations.